refactor(error) CustomError 파일 디렉토리 변경
src root에서 src/utils로 이동
